refactor to Global Controller
