--- src/main/java/zos/shell/controller/factory/type/ContainerType.java
@@ -1,40 +1,0 @@
-package zos.shell.controller.factory.type;
-
-public class ContainerType {
-
-    public enum Name {
-
-        BROWSE_JOB,
-        CANCEL,
-        CHANGE_CONNECTION,
-        CHANGE_DIRECTORY,
-        CHANGE_WINDOW,
-        CONCAT,
-        CONSOLE,
-        COPY,
-        COUNT,
-        DELETE,
-        DOWNLOAD,
-        DOWNLOAD_JOB,
-        ECHO,
-        EDIT,
-        GREP,
-        LIST,
-        LOCAL_FILE,
-        MAKE_DIR,
-        PROCESS_LIST,
-        PURGE,
-        RENAME,
-        SAVE,
-        SEARCH_CACHE,
-        STOP,
-        SUBMIT,
-        TAIL,
-        TOUCH,
-        TSO,
-        UNAME,
-        USS
-
-    }
-
-}
