+++ src/main/java/zos/shell/controller/factory/type/GlobalControllerType.java
@@ -6,8 +6,10 @@ public class GlobalControllerType {
 
         ECHO,
         ENV,
-        PATH,
         LOCAL_FILE,
+        PATH,
+        SEARCH_CACHE,
+        UNAME
 
     }
